Initial research annotation
The annotation in its original version is accepted by the Faculty Council of the Faculty of Management at the Technical University - Sofia in October 2015. As of October 2016, the PhD Student has finalized the literature review on the theory of critical success factors (CSF), stakeholder management (as a key area of modern project management), and their overall significance when it comes to open source projects, hence the focus in 2017 on conducting a rigorous empirical research to outline and analyze the set of CSF as per the dissertation’s aim.